lwb 4.16

+++ app/src/main/java/com/example/pdr_locator/algorithm/IAlgorithm.java
@@ -15,5 +15,5 @@ public interface IAlgorithm {
      * @param input 输入IMU数据
      * @return 返回一次定位结果（x,y,z）
      */
-    double[] getCoordinate(double[][] input);
+    double[] getCoordinate(float[][] input);
 }
